Do not use implicit document
The implicit document variable might not correspond to the correct
document for either parsing the source markup or rendering the
preview markup. Replace it with ownerDocument in the context.

Use case: use of pagedjs where the source markup is in a source iframe
and the rendered markup in another iframe. The implicit document
corresponds to neither

When source and target documents differs, the previewer now does not
remove the source styles